pyproject.toml
[project]
name = "cortexia-video"
version = "0.1.0"
description = "Video annotation framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "PyYAML>=6.0",
    "toml>=0.10.2",
    "pydantic>=2.0",
    "opencv-python>=4.5.0",
    "numpy>=1.21.0",
    "transformers",
    "torch",
    "pillow",
    "accelerate",
    "pretty-errors>=1.2.25",
    "modelscope>=1.25.0",
    "pyvips>=3.0.0",
    "sahi>=0.11.23",
    "matplotlib>=3.10.3",
    "sentencepiece>=0.2.0",
    "protobuf>=6.30.2",
    "decord>=0.6.0",
    "ftfy>=6.3.1",
    "einops>=0.8.1",
    "timm>=1.0.15",
    "wandb>=0.19.11",
    "tiktoken>=0.9.0",
    "fairscale>=0.4.13",
    "scipy>=1.15.3",
    "nltk>=3.9.1",
    "ipykernel>=6.29.5",
    "json-repair>=0.47.7",
    "python-dotenv>=1.1.1",
    "pylance>=0.33.0",
]

[project.scripts]
cortexia-video = "cortexia_video.cli.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["cortexia_video"]
